View영역의 DB INSERT에서 BO 메서드의 temperature가 오류난 이유 : View영역의 addSeller 메서드에서 파라미터 타입이 맞지 않아서 발생한 거였음.

+++ src/main/java/com/quiz/lesson04/mapper/SellerMapper.java
@@ -21,6 +21,6 @@ public interface SellerMapper {
 			// 때문에 하나의 Map으로 구성해야 하는데, @Param을 사용해 맵으로 생성이 가능.
 			@Param("nickname") String nickname,
 			@Param("profileImageUrl") String profileImageUrl,
-			@Param("temperature") double temperature);
+			@Param("temperature") float temperature);
 	
 }
